STAR/requirements.txt
# StarChan Discord Bot Requirements
# Install with: pip install -r requirements.txt

discord.py>=2.3.0
aiofiles>=23.0.0
